// planning-idsdp-pps-frontend/src/app/main/modules/dpp-tapp-management/models/dppProjectManagement.model.ts
export class DppProjectManagementModel {
    id: string;
    uuid: string;
    projectConceptMasterId: number;
    dppMasterId: number;
    implementationArrangement: string;
    revenueBudget: string;
    revenueBudget2: string;
    projectConceptUuid: string;
    isTransferableToRevenueBudget: boolean;
}